s/unsigned/unsigned long/ (64bit fixes, from damir@altlinux)
git-svn-id: file:///.data/svn_convert/svn@5789 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ trunk/sim/sim/plugins.cpp
@@ -580,7 +580,7 @@ void PluginManagerPrivate::saveState()
     }
 }
 
-const unsigned NO_PLUGIN = (unsigned)(-1);
+const unsigned long NO_PLUGIN = (unsigned long)(-1);
 
 void PluginManagerPrivate::reloadState()
 {
